+++ doc/changelog.md
@@ -4,8 +4,9 @@
 
 New features:
 
-- ([#23](../../../issues/23)) add container linking
-  ([example](../src/test/java/pl/domzal/junit/docker/rule/examples/ExampleLinkTest.java))
+- add static ([#23](../../../issues/23)) and dynamic ([#27](../../../issues/27)) container linking
+  ([static](../src/test/java/pl/domzal/junit/docker/rule/examples/ExampleLinkTest.java) and
+  [dynamic](../src/test/java/pl/domzal/junit/docker/rule/examples/ExampleLinkDynamicTest.java) example)
 - ([#24](../../../issues/24)) add possibility to wait for specific sequence of messages
   (instead single message) in output at container start
   ([example](../src/test/java/pl/domzal/junit/docker/rule/examples/ExampleWaitForLogMessageSequenceAtStartTest.java))